Add map lookup for keys
While this takes a little additional memory (a <string, size_t> map per
object), it scales much better when inserting a lot of hash keys, especially
when they're unique.

At some point the combination of keys and key vector could be combined into a
single data structure, but for now this improvement spends a little memory to
save a lot of time.